created room demo (multithreaded timer demo)
this is a little rough.
theres still a lot of commented code that will be removed later, and rooms cannot be deleted after creation yet

also worth noting that this disables httpOnly on the jetty JSESSIONID cookie in order to make it work. it will probably be the same for the whole project, but may be a security risk (as we aren't holding any critical data i'm not too concerned)